Extract "Hosts" from host.go
